Did some debugging on menus for when there are multiple characters in the party, fixed a couple of things and made some aesthetic improvements.
Additionally, added a rudimentary implementation of the overworld (currently with two map icons that lead to identical locations, just for starts), connected it to the other menus and debugged a problem with saving rooted in some transient variables for visual content in character sprite panes. Saving appears to work properly now.